refactor(ai-sdk): split the Google provider into providers/google/
Move the grab-bag providers/google.ts into a directory split by concern
(chat, embeddings, images, files, vector-store, filters, config), with
providers/google.ts kept as a re-export barrel so every import path is
unchanged. Hoist the duplicated sleep() helper to util/sleep.ts.

Pure code movement, no behavior change. The per-adapter getClient()
copies are moved as-is; deduping them is #972.

Refs #973
