+++ packages/ai-sdk/src/providers/google/config.ts
@@ -1,0 +1,3 @@
+export interface GoogleConfig {
+  apiKey: string
+}
