+++ packages/ai-sdk/src/providers/google/index.ts
@@ -1,0 +1,11 @@
+export type { GoogleConfig } from './config.js'
+export { GoogleProvider } from './provider.js'
+export { GoogleAdapter, mapGeminiFinishReason, toGeminiContents } from './chat.js'
+export { filterToGeminiString } from './filters.js'
+export {
+  fromGeminiFileSearchStore,
+  fromGeminiDocument,
+  attributesToCustomMetadata,
+  customMetadataToAttributes,
+  mimeTypeFromFilename,
+} from './vector-store.js'
